+++ fortscale/fortscale-collection/src/main/java/fortscale/collection/jobs/activity/UserActivitySourceDevicesHandler.java
@@ -41,7 +41,7 @@ public class UserActivitySourceDevicesHandler extends UserActivityBaseHandler {
     }
 
     @Override
-    protected void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources, long currBucketStartTime, long currBucketEndTime, Map<String, Integer> additionalActivityHistogram) {
+    protected void updateAdditionalActivitySpecificDocumentInDatabase(List<String> dataSources, long currBucketStartTime, long currBucketEndTime, Map<String, Double> additionalActivityHistogram) {
         throw new UnsupportedOperationException("not yet implemented");
     }
