+++ fortscale/fortscale-domain/src/main/java/fortscale/domain/core/activities/OrganizationActivityLocationDocument.java
@@ -22,7 +22,7 @@ public class OrganizationActivityLocationDocument extends UserActivityDocument {
     private OrganizationActivityLocationDocument.Locations locations;
 
     @Override
-    public Map<String, Integer> getHistogram() {
+    public Map<String, Double> getHistogram() {
         return getLocations().getCountryHistogram();
     }
 
@@ -37,11 +37,11 @@ public class OrganizationActivityLocationDocument extends UserActivityDocument {
     }
 
     public static class Locations {
-        private Map<String, Integer> countryHistogram = new HashMap<>();
+        private Map<String, Double> countryHistogram = new HashMap<>();
 
 
         @Field(COUNTRY_HISTOGRAM_FIELD_NAME)
-        public Map<String, Integer> getCountryHistogram() {
+        public Map<String, Double> getCountryHistogram() {
             return countryHistogram;
         }
     }
